requirements.txt
Django==5.2
django-countries==7.5.1
python-dateutil==2.8.2
whitenoise==6.6.0
gunicorn==21.2.0
